Added an MIT license to the project.
No license was explicitly distributed with the project, but the creator (james_lohr) has said that it is under a similar license to OpenTK, which is MIT/X11.
http://www.opentk.com/node/2628, "The same as OpenTk: basically do whatever you like with it. However, if you have changes that you think would be beneficial for all then do suggest them I will try to get them into the next release. The advantage for you being that you won't have to hack around each new version to include the changes you would like. :)"

Considering this, I feel comfortable assigning an MIT license to the project, however if james_lohr has any issues with this than I urge him to contact me.